refactor util
