+++ src-tauri/src/core/app/mod.rs
@@ -1,0 +1,3 @@
+pub mod commands;
+pub mod helpers;
+pub mod models;
