+++ src-tauri/src/core/downloads/mod.rs
@@ -1,0 +1,6 @@
+pub mod commands;
+pub mod helpers;
+pub mod models;
+
+#[cfg(test)]
+mod tests;
