+++ src-tauri/src/core/downloads/tests.rs
@@ -1,0 +1,341 @@
+use super::helpers::*;
+use super::models::*;
+use reqwest::header::HeaderMap;
+use std::collections::HashMap;
+
+// Helper function to create a minimal proxy config for testing
+fn create_test_proxy_config(url: &str) -> ProxyConfig {
+    ProxyConfig {
+        url: url.to_string(),
+        username: None,
+        password: None,
+        no_proxy: None,
+        ignore_ssl: None,
+    }
+}
+
+#[test]
+fn test_validate_proxy_config() {
+    // Valid HTTP proxy
+    let config = ProxyConfig {
+        url: "http://proxy.example.com:8080".to_string(),
+        username: Some("user".to_string()),
+        password: Some("pass".to_string()),
+        no_proxy: Some(vec!["localhost".to_string(), "*.example.com".to_string()]),
+        ignore_ssl: Some(true),
+    };
+    assert!(validate_proxy_config(&config).is_ok());
+
+    // Valid HTTPS proxy
+    let config = ProxyConfig {
+        url: "https://proxy.example.com:8080".to_string(),
+        username: None,
+        password: None,
+        no_proxy: None,
+        ignore_ssl: None,
+    };
+    assert!(validate_proxy_config(&config).is_ok());
+
+    // Valid SOCKS5 proxy
+    let config = ProxyConfig {
+        url: "socks5://proxy.example.com:1080".to_string(),
+        username: None,
+        password: None,
+        no_proxy: None,
+        ignore_ssl: None,
+    };
+    assert!(validate_proxy_config(&config).is_ok());
+
+    // Invalid URL
+    let config = create_test_proxy_config("invalid-url");
+    assert!(validate_proxy_config(&config).is_err());
+
+    // Unsupported scheme
+    let config = create_test_proxy_config("ftp://proxy.example.com:21");
+    assert!(validate_proxy_config(&config).is_err());
+
+    // Username without password
+    let mut config = create_test_proxy_config("http://proxy.example.com:8080");
+    config.username = Some("user".to_string());
+    assert!(validate_proxy_config(&config).is_err());
+
+    // Password without username
+    let mut config = create_test_proxy_config("http://proxy.example.com:8080");
+    config.password = Some("pass".to_string());
+    assert!(validate_proxy_config(&config).is_err());
+
+    // Empty no_proxy entry
+    let mut config = create_test_proxy_config("http://proxy.example.com:8080");
+    config.no_proxy = Some(vec!["".to_string()]);
+    assert!(validate_proxy_config(&config).is_err());
+
+    // Invalid wildcard pattern
+    let mut config = create_test_proxy_config("http://proxy.example.com:8080");
+    config.no_proxy = Some(vec!["*.".to_string()]);
+    assert!(validate_proxy_config(&config).is_err());
+}
+
+#[test]
+fn test_should_bypass_proxy() {
+    let no_proxy = vec![
+        "localhost".to_string(),
+        "127.0.0.1".to_string(),
+        "*.example.com".to_string(),
+        "specific.domain.com".to_string(),
+    ];
+
+    // Should bypass for localhost
+    assert!(should_bypass_proxy("http://localhost:8080/path", &no_proxy));
+
+    // Should bypass for 127.0.0.1
+    assert!(should_bypass_proxy("https://127.0.0.1:3000/api", &no_proxy));
+
+    // Should bypass for wildcard match
+    assert!(should_bypass_proxy(
+        "http://sub.example.com/path",
+        &no_proxy
+    ));
+    assert!(should_bypass_proxy("https://api.example.com/v1", &no_proxy));
+
+    // Should bypass for specific domain
+    assert!(should_bypass_proxy(
+        "http://specific.domain.com/test",
+        &no_proxy
+    ));
+
+    // Should NOT bypass for other domains
+    assert!(!should_bypass_proxy("http://other.com/path", &no_proxy));
+    assert!(!should_bypass_proxy("https://example.org/api", &no_proxy));
+
+    // Should bypass everything with "*"
+    let wildcard_no_proxy = vec!["*".to_string()];
+    assert!(should_bypass_proxy(
+        "http://any.domain.com/path",
+        &wildcard_no_proxy
+    ));
+
+    // Empty no_proxy should not bypass anything
+    let empty_no_proxy = vec![];
+    assert!(!should_bypass_proxy(
+        "http://any.domain.com/path",
+        &empty_no_proxy
+    ));
+}
+
+#[test]
+fn test_create_proxy_from_config() {
+    // Valid configuration should work
+    let mut config = create_test_proxy_config("http://proxy.example.com:8080");
+    config.username = Some("user".to_string());
+    config.password = Some("pass".to_string());
+    assert!(create_proxy_from_config(&config).is_ok());
+
+    // Invalid configuration should fail
+    let config = create_test_proxy_config("invalid-url");
+    assert!(create_proxy_from_config(&config).is_err());
+}
+
+#[test]
+fn test_convert_headers() {
+    let mut headers = HashMap::new();
+    headers.insert("User-Agent".to_string(), "test-agent".to_string());
+    headers.insert("Authorization".to_string(), "Bearer token".to_string());
+
+    let header_map = _convert_headers(&headers).unwrap();
+    assert_eq!(header_map.len(), 2);
+    assert_eq!(header_map.get("User-Agent").unwrap(), "test-agent");
+    assert_eq!(header_map.get("Authorization").unwrap(), "Bearer token");
+}
+
+#[test]
+fn test_proxy_ssl_verification_settings() {
+    // Test proxy config with SSL verification settings
+    let mut config = create_test_proxy_config("https://proxy.example.com:8080");
+    config.ignore_ssl = Some(true);
+
+    // Should validate successfully
+    assert!(validate_proxy_config(&config).is_ok());
+
+    // Test with all SSL settings as false
+    config.ignore_ssl = Some(false);
+
+    // Should still validate successfully
+    assert!(validate_proxy_config(&config).is_ok());
+}
+
+#[test]
+fn test_proxy_config_with_mixed_ssl_settings() {
+    // Test with mixed SSL settings - ignore_ssl true, others false
+    let mut config = create_test_proxy_config("https://proxy.example.com:8080");
+    config.ignore_ssl = Some(true);
+
+    assert!(validate_proxy_config(&config).is_ok());
+    assert!(create_proxy_from_config(&config).is_ok());
+}
+
+#[test]
+fn test_proxy_config_ssl_defaults() {
+    // Test with no SSL settings (should use None defaults)
+    let config = create_test_proxy_config("https://proxy.example.com:8080");
+
+    assert_eq!(config.ignore_ssl, None);
+
+    assert!(validate_proxy_config(&config).is_ok());
+    assert!(create_proxy_from_config(&config).is_ok());
+}
+
+#[test]
+fn test_download_item_with_ssl_proxy() {
+    // Test that DownloadItem can be created with SSL proxy configuration
+    let mut proxy_config = create_test_proxy_config("https://proxy.example.com:8080");
+    proxy_config.ignore_ssl = Some(true);
+
+    let download_item = DownloadItem {
+        url: "https://example.com/file.zip".to_string(),
+        save_path: "downloads/file.zip".to_string(),
+        proxy: Some(proxy_config),
+    };
+
+    assert!(download_item.proxy.is_some());
+    let proxy = download_item.proxy.unwrap();
+    assert_eq!(proxy.ignore_ssl, Some(true));
+}
+
+#[test]
+fn test_client_creation_with_ssl_settings() {
+    // Test client creation with SSL settings
+    let mut proxy_config = create_test_proxy_config("https://proxy.example.com:8080");
+    proxy_config.ignore_ssl = Some(true);
+
+    let download_item = DownloadItem {
+        url: "https://example.com/file.zip".to_string(),
+        save_path: "downloads/file.zip".to_string(),
+        proxy: Some(proxy_config),
+    };
+
+    let header_map = HeaderMap::new();
+    let result = _get_client_for_item(&download_item, &header_map);
+
+    // Should create client successfully even with SSL settings
+    assert!(result.is_ok());
+}
+
+#[test]
+fn test_proxy_config_with_http_and_ssl_settings() {
+    // Test that SSL settings work with HTTP proxy (though not typically used)
+    let mut config = create_test_proxy_config("http://proxy.example.com:8080");
+    config.ignore_ssl = Some(true);
+
+    assert!(validate_proxy_config(&config).is_ok());
+    assert!(create_proxy_from_config(&config).is_ok());
+}
+
+#[test]
+fn test_proxy_config_with_socks_and_ssl_settings() {
+    // Test that SSL settings work with SOCKS proxy
+    let mut config = create_test_proxy_config("socks5://proxy.example.com:1080");
+    config.ignore_ssl = Some(false);
+
+    assert!(validate_proxy_config(&config).is_ok());
+
+    // SOCKS proxies are not supported by reqwest::Proxy::all()
+    // This test should expect an error for SOCKS proxies
+    let result = create_proxy_from_config(&config);
+    assert!(result.is_err());
+
+    // Test with HTTP proxy instead which is supported
+    let mut http_config = create_test_proxy_config("http://proxy.example.com:8080");
+    http_config.ignore_ssl = Some(false);
+    assert!(validate_proxy_config(&http_config).is_ok());
+    assert!(create_proxy_from_config(&http_config).is_ok());
+}
+
+#[test]
+fn test_download_item_creation() {
+    let item = DownloadItem {
+        url: "https://example.com/file.tar.gz".to_string(),
+        save_path: "models/test.tar.gz".to_string(),
+        proxy: None,
+    };
+
+    assert_eq!(item.url, "https://example.com/file.tar.gz");
+    assert_eq!(item.save_path, "models/test.tar.gz");
+}
+
+#[test]
+fn test_download_event_creation() {
+    let event = DownloadEvent {
+        transferred: 1024,
+        total: 2048,
+    };
+
+    assert_eq!(event.transferred, 1024);
+    assert_eq!(event.total, 2048);
+}
+
+#[test]
+fn test_err_to_string() {
+    let error = "Test error";
+    let result = err_to_string(error);
+    assert_eq!(result, "Error: Test error");
+}
+
+#[test]
+fn test_convert_headers_valid() {
+    let mut headers = HashMap::new();
+    headers.insert("Content-Type".to_string(), "application/json".to_string());
+    headers.insert("Authorization".to_string(), "Bearer token123".to_string());
+
+    let result = _convert_headers(&headers);
+    assert!(result.is_ok());
+
+    let header_map = result.unwrap();
+    assert_eq!(header_map.len(), 2);
+    assert_eq!(header_map.get("Content-Type").unwrap(), "application/json");
+    assert_eq!(header_map.get("Authorization").unwrap(), "Bearer token123");
+}
+
+#[test]
+fn test_convert_headers_invalid_header_name() {
+    let mut headers = HashMap::new();
+    headers.insert("Invalid\nHeader".to_string(), "value".to_string());
+
+    let result = _convert_headers(&headers);
+    assert!(result.is_err());
+}
+
+#[test]
+fn test_convert_headers_invalid_header_value() {
+    let mut headers = HashMap::new();
+    headers.insert("Content-Type".to_string(), "invalid\nvalue".to_string());
+
+    let result = _convert_headers(&headers);
+    assert!(result.is_err());
+}
+
+#[test]
+fn test_download_manager_state_default() {
+    let state = DownloadManagerState::default();
+    assert!(state.cancel_tokens.is_empty());
+}
+
+#[test]
+fn test_download_event_serialization() {
+    let event = DownloadEvent {
+        transferred: 512,
+        total: 1024,
+    };
+
+    let json = serde_json::to_string(&event).unwrap();
+    assert!(json.contains("\"transferred\":512"));
+    assert!(json.contains("\"total\":1024"));
+}
+
+#[test]
+fn test_download_item_deserialization() {
+    let json = r#"{"url":"https://example.com/file.zip","save_path":"downloads/file.zip"}"#;
+    let item: DownloadItem = serde_json::from_str(json).unwrap();
+
+    assert_eq!(item.url, "https://example.com/file.zip");
+    assert_eq!(item.save_path, "downloads/file.zip");
+}
