+++ src-tauri/src/core/extensions/mod.rs
@@ -1,0 +1,1 @@
+pub mod commands;
