+++ src-tauri/src/core/filesystem/helpers.rs
@@ -1,0 +1,23 @@
+use crate::core::app::commands::get_jan_data_folder_path;
+use jan_utils::normalize_file_path;
+use std::path::PathBuf;
+use tauri::Runtime;
+
+pub fn resolve_path<R: Runtime>(app_handle: tauri::AppHandle<R>, path: &str) -> PathBuf {
+    let path = if path.starts_with("file:/") || path.starts_with("file:\\") {
+        let normalized = normalize_file_path(path);
+        let relative_normalized = normalized
+            .trim_start_matches(std::path::MAIN_SEPARATOR)
+            .trim_start_matches('/')
+            .trim_start_matches('\\');
+        get_jan_data_folder_path(app_handle).join(relative_normalized)
+    } else {
+        PathBuf::from(path)
+    };
+
+    if path.starts_with("http://") || path.starts_with("https://") {
+        path
+    } else {
+        path.canonicalize().unwrap_or(path)
+    }
+}
