+++ src-tauri/src/core/filesystem/tests.rs
@@ -1,0 +1,90 @@
+use super::commands::*;
+use crate::core::app::commands::get_jan_data_folder_path;
+use std::fs::{self, File};
+use std::io::Write;
+use tauri::test::mock_app;
+
+#[test]
+fn test_rm() {
+    let app = mock_app();
+    let path = "test_rm_dir";
+    fs::create_dir_all(get_jan_data_folder_path(app.handle().clone()).join(path)).unwrap();
+    let args = vec![format!("file://{}", path).to_string()];
+    let result = rm(app.handle().clone(), args);
+    assert!(result.is_ok());
+    assert!(!get_jan_data_folder_path(app.handle().clone())
+        .join(path)
+        .exists());
+}
+
+#[test]
+fn test_mkdir() {
+    let app = mock_app();
+    let path = "test_mkdir_dir";
+    let args = vec![format!("file://{}", path).to_string()];
+    let result = mkdir(app.handle().clone(), args);
+    assert!(result.is_ok());
+    assert!(get_jan_data_folder_path(app.handle().clone())
+        .join(path)
+        .exists());
+    let _ = fs::remove_dir_all(get_jan_data_folder_path(app.handle().clone()).join(path));
+}
+
+#[test]
+fn test_join_path() {
+    let app = mock_app();
+    let path = "file://test_dir";
+    let args = vec![path.to_string(), "test_file".to_string()];
+    let result = join_path(app.handle().clone(), args).unwrap();
+    assert_eq!(
+        result,
+        get_jan_data_folder_path(app.handle().clone())
+            .join(&format!("test_dir{}test_file", std::path::MAIN_SEPARATOR))
+            .to_string_lossy()
+            .to_string()
+    );
+}
+
+#[test]
+fn test_exists_sync() {
+    let app = mock_app();
+    let path = "file://test_exists_sync_file";
+    let dir_path = get_jan_data_folder_path(app.handle().clone());
+    fs::create_dir_all(&dir_path).unwrap();
+    let file_path = dir_path.join("test_exists_sync_file");
+    File::create(&file_path).unwrap();
+    let args: Vec<String> = vec![path.to_string()];
+    let result = exists_sync(app.handle().clone(), args).unwrap();
+    assert!(result);
+    fs::remove_file(file_path).unwrap();
+}
+
+#[test]
+fn test_read_file_sync() {
+    let app = mock_app();
+    let path = "file://test_read_file_sync_file";
+    let dir_path = get_jan_data_folder_path(app.handle().clone());
+    fs::create_dir_all(&dir_path).unwrap();
+    let file_path = dir_path.join("test_read_file_sync_file");
+    let mut file = File::create(&file_path).unwrap();
+    file.write_all(b"test content").unwrap();
+    let args = vec![path.to_string()];
+    let result = read_file_sync(app.handle().clone(), args).unwrap();
+    assert_eq!(result, "test content".to_string());
+    fs::remove_file(file_path).unwrap();
+}
+
+#[test]
+fn test_readdir_sync() {
+    let app = mock_app();
+    let dir_path = get_jan_data_folder_path(app.handle().clone()).join("test_readdir_sync_dir");
+    fs::create_dir_all(&dir_path).unwrap();
+    File::create(dir_path.join("file1.txt")).unwrap();
+    File::create(dir_path.join("file2.txt")).unwrap();
+
+    let args = vec![dir_path.to_string_lossy().to_string()];
+    let result = readdir_sync(app.handle().clone(), args).unwrap();
+    assert_eq!(result.len(), 2);
+
+    let _ = fs::remove_dir_all(dir_path);
+}
