+++ src-tauri/src/core/mcp/mod.rs
@@ -1,0 +1,5 @@
+pub mod commands;
+pub mod helpers;
+
+#[cfg(test)]
+mod tests;
