+++ src-tauri/src/core/mod.rs
@@ -1,9 +1,10 @@
-pub mod cmd;
-pub mod download;
-pub mod fs;
+pub mod app;
+pub mod downloads;
+pub mod extensions;
+pub mod filesystem;
 pub mod mcp;
 pub mod server;
 pub mod setup;
 pub mod state;
+pub mod system;
 pub mod threads;
-pub mod threads_utils;
