+++ src-tauri/src/core/setup.rs
@@ -13,9 +13,8 @@ use tauri_plugin_store::StoreExt;
 
 // MCP
 use super::{
-    cmd::{get_jan_data_folder_path, get_jan_extensions_path},
-    mcp::run_mcp_commands,
-    state::AppState,
+    app::commands::get_jan_data_folder_path, extensions::commands::get_jan_extensions_path,
+    mcp::helpers::run_mcp_commands, mcp::helpers::stop_mcp_servers, state::AppState,
 };
 
 pub fn install_extensions(app: tauri::AppHandle, force: bool) -> Result<(), String> {
@@ -165,7 +164,7 @@ pub fn install_extensions(app: tauri::AppHandle, force: bool) -> Result<(), Stri
     Ok(())
 }
 
-fn extract_extension_manifest<R: Read>(
+pub fn extract_extension_manifest<R: Read>(
     archive: &mut Archive<R>,
 ) -> Result<Option<serde_json::Value>, String> {
     let entry = archive
@@ -207,7 +206,7 @@ pub fn setup_mcp(app: &App) {
             log::info!("Received kill-mcp-servers event - cleaning up MCP servers");
             let app_state = app_handle.state::<AppState>();
             // Stop all running MCP servers
-            if let Err(e) = super::mcp::stop_mcp_servers(app_state.mcp_servers.clone()).await {
+            if let Err(e) = stop_mcp_servers(app_state.mcp_servers.clone()).await {
                 log::error!("Failed to stop MCP servers: {}", e);
                 return;
             }
