+++ src-tauri/src/core/system/mod.rs
@@ -1,0 +1,1 @@
+pub mod commands;
