+++ src-tauri/src/lib.rs
@@ -1,7 +1,7 @@
 mod core;
 use core::{
-    cmd::get_jan_data_folder_path,
-    download::DownloadManagerState,
+    app::commands::get_jan_data_folder_path,
+    downloads::models::DownloadManagerState,
     setup::{self, setup_mcp},
     state::AppState,
 };
@@ -35,62 +35,66 @@ pub fn run() {
         .plugin(tauri_plugin_hardware::init())
         .invoke_handler(tauri::generate_handler![
             // FS commands - Deperecate soon
-            core::fs::join_path,
-            core::fs::mkdir,
-            core::fs::exists_sync,
-            core::fs::readdir_sync,
-            core::fs::read_file_sync,
-            core::fs::rm,
-            core::fs::file_stat,
-            core::fs::write_file_sync,
-            // App commands
-            core::cmd::get_app_configurations,
-            core::cmd::get_active_extensions,
-            core::cmd::get_user_home_path,
-            core::cmd::update_app_configuration,
-            core::cmd::get_jan_data_folder_path,
-            core::cmd::get_jan_extensions_path,
-            core::cmd::relaunch,
-            core::cmd::open_app_directory,
-            core::cmd::open_file_explorer,
-            core::cmd::install_extensions,
-            core::cmd::app_token,
-            core::cmd::start_server,
-            core::cmd::stop_server,
-            core::cmd::get_server_status,
-            core::cmd::read_logs,
-            core::cmd::change_app_data_folder,
-            core::cmd::factory_reset,
+            core::filesystem::commands::join_path,
+            core::filesystem::commands::mkdir,
+            core::filesystem::commands::exists_sync,
+            core::filesystem::commands::readdir_sync,
+            core::filesystem::commands::read_file_sync,
+            core::filesystem::commands::rm,
+            core::filesystem::commands::file_stat,
+            core::filesystem::commands::write_file_sync,
+            core::filesystem::commands::write_yaml,
+            core::filesystem::commands::read_yaml,
+            core::filesystem::commands::decompress,
+            // App configuration commands
+            core::app::commands::get_app_configurations,
+            core::app::commands::get_user_home_path,
+            core::app::commands::update_app_configuration,
+            core::app::commands::get_jan_data_folder_path,
+            core::app::commands::get_configuration_file_path,
+            core::app::commands::default_data_folder_path,
+            core::app::commands::change_app_data_folder,
+            core::app::commands::app_token,
+            // Extension commands
+            core::extensions::commands::get_jan_extensions_path,
+            core::extensions::commands::install_extensions,
+            core::extensions::commands::get_active_extensions,
+            // System commands
+            core::system::commands::relaunch,
+            core::system::commands::open_app_directory,
+            core::system::commands::open_file_explorer,
+            core::system::commands::factory_reset,
+            core::system::commands::read_logs,
+            core::system::commands::is_library_available,
+            // Server commands
+            core::server::commands::start_server,
+            core::server::commands::stop_server,
+            core::server::commands::get_server_status,
             // MCP commands
-            core::mcp::get_tools,
-            core::mcp::call_tool,
-            core::mcp::restart_mcp_servers,
-            core::mcp::get_connected_servers,
-            core::mcp::save_mcp_configs,
-            core::mcp::get_mcp_configs,
-            core::mcp::activate_mcp_server,
-            core::mcp::deactivate_mcp_server,
-            core::mcp::reset_mcp_restart_count,
+            core::mcp::commands::get_tools,
+            core::mcp::commands::call_tool,
+            core::mcp::commands::restart_mcp_servers,
+            core::mcp::commands::get_connected_servers,
+            core::mcp::commands::save_mcp_configs,
+            core::mcp::commands::get_mcp_configs,
+            core::mcp::commands::activate_mcp_server,
+            core::mcp::commands::deactivate_mcp_server,
+            core::mcp::commands::reset_mcp_restart_count,
             // Threads
-            core::threads::list_threads,
-            core::threads::create_thread,
-            core::threads::modify_thread,
-            core::threads::delete_thread,
-            core::threads::list_messages,
-            core::threads::create_message,
-            core::threads::modify_message,
-            core::threads::delete_message,
-            core::threads::get_thread_assistant,
-            core::threads::create_thread_assistant,
-            core::threads::modify_thread_assistant,
-            // threads utils
-            core::threads_utils::write_yaml,
-            core::threads_utils::read_yaml,
-            core::threads_utils::decompress,
-            core::threads_utils::is_library_available,
+            core::threads::commands::list_threads,
+            core::threads::commands::create_thread,
+            core::threads::commands::modify_thread,
+            core::threads::commands::delete_thread,
+            core::threads::commands::list_messages,
+            core::threads::commands::create_message,
+            core::threads::commands::modify_message,
+            core::threads::commands::delete_message,
+            core::threads::commands::get_thread_assistant,
+            core::threads::commands::create_thread_assistant,
+            core::threads::commands::modify_thread_assistant,
             // Download
-            core::download::download_files,
-            core::download::cancel_download_task,
+            core::downloads::commands::download_files,
+            core::downloads::commands::cancel_download_task,
         ])
         .manage(AppState {
             app_token: Some(generate_app_token()),
